feat: Iniciando branch dev, alterado projeto de ts para js
