+++ cypress/e2e/submissao-proposta.cy.js
@@ -1,0 +1,25 @@
+describe('Sistema Integrado de Gestão para Fundações de Amparo a Pesquisas', () => {
+  beforeEach(() => {
+    // Gancho em nível raíz
+    // executa antes de realizar cada teste(it)
+    cy.typelogin(
+      '[URL do sistema]',// [URL do sistema]
+      '[E-mail do usuário]', // [E-mail do usuário]
+      '[Senha do usuário]', // [Senha do usuário]
+    ); //Acessa a página de login usando as credenciais do usuário e senha.
+  });
+  it('Realiza login no sistema e submete uma proposta', () => {
+    cy.get('[data-cy="breadcrumb-home"]').click(); //Clica no botão "Home" para retornar à página anterior
+    cy.get('[data-cy="editais-ver-mais"]').click(); //Clica no botão "Ver Mais" para acessar a página de Editais
+
+    cy.get('[data-cy="visualizar-edital-grupo-01-e.s.-0"]').click(); //Edite essa linha para selecionar o Edital respectivo
+
+    cy.wait(300); //Aguarda 300ms para garantir que a página foi carregada completamente
+    cy.get('[data-cy="criar-proposta"]').click(); //Clica no botão "Criar Proposta" para iniciar o processo de criação de uma nova proposta
+    cy.get('[data-cy="tituloDoProjeto"]').type(
+        'Submissão de Proposta Cypress', //Preenche o campo "Título do Projeto" com o valor "Submissão de Proposta de Teste"
+        { delay: 0 },
+    )
+    //Atividade 3 - Faça a continuidade do teste, preenchendo os campos obrigatórios da proposta.
+  }); 
+});
