+++ cypress/helpers/date.helper.js
@@ -1,0 +1,19 @@
+export const getCurrentDateTime = (
+  addSeconds,
+  addMinutes,
+  addHours,
+  addDays,
+  addMonths,
+  addYears)=> {
+  const now = new Date();
+  if (addSeconds) now.setSeconds(now.getSeconds() + addSeconds);
+  if (addMinutes) now.setMinutes(now.getMinutes() + addMinutes);
+  if (addHours) now.setHours(now.getHours() + addHours);
+  if (addDays) now.setDate(now.getDate() + addDays);
+  if (addMonths) now.setMonth(now.getMonth() + addMonths);
+  if (addYears) now.setFullYear(now.getFullYear() + addYears);
+  const date = now.toLocaleDateString('pt-BR'); // "08/05/2025"
+  const time = now.toLocaleTimeString('pt-BR', { hour12: false }); // "16:30:09"
+  const [hh, mm, ss] = time.split(':');
+  return `${date} ${mm}.${ss}.${hh}`;
+};
