--- cypress/support/index.ts
@@ -1,5 +1,0 @@
-declare namespace Cypress{
-    interface Chainable{
-        typelogin: (url: string, email: string, password: string) => void
-    }
-}
