+++ env.config/test.json
@@ -1,0 +1,9 @@
+{
+  "baseUrl": "https://novo-sig.ledes.net/",
+  "specPattern": "cypress/e2e/**/**.js",
+  "env": {
+    "username": "grupo6_gestor@sig.com",
+    "password": "Grupo6@sig",
+    "envName": "test"
+  }
+}
